ouai les provlèmes

+++ include/neural_network/neural_struct.h
@@ -26,7 +26,7 @@ typedef struct network
 typedef struct bot
 {
     network *net;
-    float score;
+    double score;
 }bot;
 
 typedef struct generation
